Complete renaming so the code is easier to read
