+++ src/eva/user_simulator/__init__.py
@@ -1,7 +1,7 @@
 """Pluggable user simulator providers."""
 
 from eva.user_simulator.base import AbstractUserSimulator
-from eva.user_simulator.client import UserSimulator
+from eva.user_simulator.elevenlabs import ElevenLabsUserSimulator
 from eva.user_simulator.factory import create_user_simulator
 
-__all__ = ["AbstractUserSimulator", "UserSimulator", "create_user_simulator"]
+__all__ = ["AbstractUserSimulator", "ElevenLabsUserSimulator", "create_user_simulator"]
