+++ tests/unit/user_simulator/test_elevenlabs.py
@@ -1,4 +1,4 @@
-"""Unit tests for UserSimulator client.
+"""Unit tests for ElevenLabsUserSimulator.
 
 Focuses on non-trivial logic: conversation end idempotency, keep-alive
 inactivity detection, end_call API polling with backoff, and error handling.
@@ -11,11 +11,11 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from eva.user_simulator.client import UserSimulator
+from eva.user_simulator.elevenlabs import ElevenLabsUserSimulator
 
 
-def _make_simulator(tmp_path: Path, **overrides) -> UserSimulator:
-    """Create a UserSimulator with minimal config for testing."""
+def _make_simulator(tmp_path: Path, **overrides) -> ElevenLabsUserSimulator:
+    """Create an ElevenLabsUserSimulator with minimal config for testing."""
     defaults = {
         "current_date_time": "2026-03-23 10:00:00",
         "persona_config": {
@@ -42,7 +42,7 @@ def _make_simulator(tmp_path: Path, **overrides) -> UserSimulator:
         "timeout": 60,
     }
     defaults.update(overrides)
-    return UserSimulator(**defaults)
+    return ElevenLabsUserSimulator(**defaults)
 
 
 def _make_conv_details(transcript=None, status="done"):
@@ -162,7 +162,7 @@ class TestCheckEndCallViaApi:
         details = _make_conv_details(transcript=[turn_with_end_call])
         sim._client.conversational_ai.conversations.get.return_value = details
 
-        with patch("eva.user_simulator.client.asyncio.sleep", new_callable=AsyncMock):
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", new_callable=AsyncMock):
             result = await sim._check_end_call_via_api("conv-123")
 
         assert result is True
@@ -178,7 +178,7 @@ class TestCheckEndCallViaApi:
         details = _make_conv_details(transcript=[turn])
         sim._client.conversational_ai.conversations.get.return_value = details
 
-        with patch("eva.user_simulator.client.asyncio.sleep", new_callable=AsyncMock):
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", new_callable=AsyncMock):
             result = await sim._check_end_call_via_api("conv-123")
 
         assert result is False
@@ -193,7 +193,7 @@ class TestCheckEndCallViaApi:
         details = _make_conv_details(transcript=[turn])
         sim._client.conversational_ai.conversations.get.return_value = details
 
-        with patch("eva.user_simulator.client.asyncio.sleep", new_callable=AsyncMock):
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", new_callable=AsyncMock):
             result = await sim._check_end_call_via_api("conv-123")
 
         assert result is False
@@ -220,7 +220,7 @@ class TestCheckEndCallViaApi:
         async def track_sleep(duration):
             sleep_delays.append(duration)
 
-        with patch("eva.user_simulator.client.asyncio.sleep", side_effect=track_sleep):
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", side_effect=track_sleep):
             result = await sim._check_end_call_via_api("conv-123")
 
         assert result is True
@@ -237,7 +237,7 @@ class TestCheckEndCallViaApi:
         empty_details = _make_conv_details(transcript=None, status="in-progress")
         sim._client.conversational_ai.conversations.get.return_value = empty_details
 
-        with patch("eva.user_simulator.client.asyncio.sleep", new_callable=AsyncMock):
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", new_callable=AsyncMock):
             result = await sim._check_end_call_via_api("conv-123")
 
         assert result is False
@@ -258,7 +258,7 @@ class TestCheckEndCallViaApi:
         async def track_sleep(duration):
             sleep_delays.append(duration)
 
-        with patch("eva.user_simulator.client.asyncio.sleep", side_effect=track_sleep):
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", side_effect=track_sleep):
             await sim._check_end_call_via_api("conv-123")
 
         # Delays: 2.0, 4.0, 8.0, 10.0 (capped), 10.0 (capped)
@@ -274,7 +274,7 @@ class TestCheckEndCallViaApi:
         details = _make_conv_details(transcript=[turn])
         sim._client.conversational_ai.conversations.get.return_value = details
 
-        with patch("eva.user_simulator.client.asyncio.sleep", new_callable=AsyncMock):
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", new_callable=AsyncMock):
             await sim._check_end_call_via_api("conv-123")
 
         details_path = tmp_path / "elevenlabs_conversation_details.json"
@@ -303,8 +303,8 @@ class TestKeepAliveTask:
         async def fake_run_in_executor(*args):
             return None
 
-        with patch("eva.user_simulator.client.asyncio.sleep", side_effect=fake_sleep):
-            with patch("eva.user_simulator.client.asyncio.get_event_loop") as mock_loop:
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", side_effect=fake_sleep):
+            with patch("eva.user_simulator.elevenlabs.asyncio.get_event_loop") as mock_loop:
                 mock_loop.return_value.run_in_executor = fake_run_in_executor
                 await sim._keep_alive_task()
 
@@ -339,8 +339,8 @@ class TestKeepAliveTask:
         async def fake_run_in_executor(*args):
             return None
 
-        with patch("eva.user_simulator.client.asyncio.sleep", side_effect=fake_sleep):
-            with patch("eva.user_simulator.client.asyncio.get_event_loop") as mock_loop:
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", side_effect=fake_sleep):
+            with patch("eva.user_simulator.elevenlabs.asyncio.get_event_loop") as mock_loop:
                 mock_loop.return_value.run_in_executor = fake_run_in_executor
                 await sim._keep_alive_task()
 
@@ -352,7 +352,7 @@ class TestKeepAliveTask:
         sim = _make_simulator(tmp_path)
         sim._conversation = MagicMock()
 
-        with patch("eva.user_simulator.client.asyncio.sleep", side_effect=asyncio.CancelledError):
+        with patch("eva.user_simulator.elevenlabs.asyncio.sleep", side_effect=asyncio.CancelledError):
             with pytest.raises(asyncio.CancelledError):
                 await sim._keep_alive_task()
